+++ src/main/kotlin/no/nav/helse/fritakagp/web/FritakModule.kt
@@ -40,7 +40,6 @@ fun Application.fritakModule(config: ApplicationConfig = environment.config) {
                 kroniskRoutes(get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get())
                 gravidRoutes(get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get())
                 altinnRoutes(get())
-                virksomhetsRoutes(get(), get(), get())
             }
         }
         swaggerRoutes("$apiBasePath")
